tickets seller/requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23